+++ service/controller/resource/tccpn/template/params_main_outputs.go
@@ -4,4 +4,5 @@ type ParamsMainOutputs struct {
 	InstanceType    string
 	MasterReplicas  int
 	OperatorVersion string
+	ReleaseVersion  string
 }