+++ service/controller/resource/tccpn/template/template_main_outputs.go
@@ -8,5 +8,7 @@ const TemplateMainOutputs = `
     Value: {{ .Outputs.MasterReplicas }}
   OperatorVersion:
     Value: {{ .Outputs.OperatorVersion }}
+  ReleaseVersion:
+    Value: {{ .Outputs.ReleaseVersion }}
 {{- end -}}
 `