+++ service/controller/resource/tcnp/template/params_main_outputs.go
@@ -4,6 +4,7 @@ type ParamsMainOutputs struct {
 	DockerVolumeSizeGB string
 	Instance           ParamsMainOutputsInstance
 	OperatorVersion    string
+	ReleaseVersion     string
 }
 
 type ParamsMainOutputsInstance struct {