+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add release version tag for ec2 instances
+- Update Cloudformation Stack when components version differ
+
+### Fixed
+
 - Fix IAM policy on Tenant Clusters to manages IAM Role tags.
 
 ## [8.7.5] - 2020-07-30